// src/types/bond.ts
// Основной интерфейс облигации на основе данных с dohod.ru
export interface Bond {
	id: string;
	xml_isin: string;
	name: string;
	emitent: string;
	borrower: string;
	borrower_country: string;
	issue_date: string;
	maturity_date: string;
	effective_date: string;
	years_to_maturity: string;
	nominal: string;
	coupon: string;
	coupon_period: string;
	price: string;
	nkd: string;
	tax_free: string;
	tax_currency_free: string;
	tax_longterm_free: string;
	isiis: string;
	rutrade: string;
	minlot: string;
	minlottype: string;
	price_return: string;
	total_return: string | null;
	mod_duration: string;
	credit_rating: string;
	credit_rating_text: string;
	description: string | null;
	comment: string | null;
	instrument_type: string;
	emitent_type: string;
	bond_type: string;
	coupon_type: string;
	moody: string | null;
	sp: string | null;
	akra: string | null;
	fitch: string | null;
	expert: string | null;
	down_risk: string;
	quality: string;
	is_qualify: string;
	size: string;
	complexity: string;
	liquidity_xml: string;
	currency: string;
	sector_text: string;
	is_show: string;
	strategy1: string;
	strategy2: string;
	strategy3: string;
	strategy4: string;
	strategy5: string;
	strategy6: string;
	event: string;
	category: string;
	vol_median: string;
	bond_link: string;
	company_id: string | null;
	q_out: string;
	q_in: string;
	q_dp1: string | null;
	q_ratingroe: string;
	q_ratingnetdebt_equity: string;
	q_col: string;
	q_outdp1: string | null;
	q_valueroe: string | null;
	q_valuenetdebt_equity: string;
	q_valuecol: string | null;
	q_balance: string;
	q_earnings: string;
	q_indp1: string | null;
	q_ros: string | null;
	q_operprof: string | null;
	q_inventorturnov: string | null;
	q_turnovofcurasset: string | null;
	q_receivableturnov: string | null;
	q_earningsdp2: string | null;
	q_balanceliq: string | null;
	q_currentliq: string | null;
	q_quiqliq: string | null;
	q_stability: string | null;
	q_shortliabilities: string | null;
	q_balancedp3: string | null;
	q_valueros: string | null;
	q_valueoperprof: string | null;
	q_valueinventorturnov: string | null;
	q_valueturnovofcurasset: string | null;
	q_valuereceivableturnov: string | null;
	q_valuecurrentliq: string | null;
	q_valuequiqliq: string | null;
	q_valueshortliabilities: string | null;
	seniority: string;
	is_roe: string;
	q_ratingcashratio: string | null;
	q_valuecashratio: string | null;
	value_of_emission: string;
	next_coupon_date: string | null;
	simple_yield: string | null;
	eqiv_simple_yield: string | null;
	reinvesting_part: string | null;
	best_score: string;
	guarantor_type: string;
	q_fear: string | null;
	bond_std: string | null;
	current_coupon_prc: string;
	last_cup: string | null;
	current_yield: string;
	eventFlag: string;
	frnIndex: string;
	frnIndexAdd: string;
	calc_isin: string;
	future_cost: string;
	not_reinvesting: string;
	reinvesting: string;
	fact_yield: string;
	calc_current_yield: string;
}